user_tab sharding prepare (finished 5 sharding of the table)

+++ internel/dao/commentReply.go
@@ -7,14 +7,14 @@ import (
 
 type CommentReplyBrief struct {
 	ReplyId       uint      `json:"reply_id"`
-	ReplyFromId   uint      `json:"reply_from_id"`
+	ReplyFromId   int64     `json:"reply_from_id"`
 	ReplyFromName string    `json:"reply_from_name"`
 	ReplyContent  string    `json:"reply_content"`
 	CreatedAt     time.Time `json:"created_at"`
 }
 
 // CreateCommentReply 创建用户回复
-func (d *Dao) CreateCommentReply(userId uint, username string, replyToId uint, replyToName string, productId uint, commentId uint, replyContent string) (uint, error) {
+func (d *Dao) CreateCommentReply(userId int64, username string, replyToId int64, replyToName string, productId uint, commentId uint, replyContent string) (uint, error) {
 	commentReply := model.CommentReply{
 		ReplyFromId:   userId,
 		ReplyFromName: username,
